- Fixed missing deletion confirmation prompts for some items. - Fixed an issue with initiation grades not properly removing themselves.

+++ Chummer/Backend/Equipment/Armor.cs
@@ -1554,8 +1554,14 @@ namespace Chummer.Backend.Equipment
         #endregion
         #endregion
 
-        public bool Remove(Character characterObject)
+        public bool Remove(Character characterObject, bool confirmDelete = true)
         {
+            if (confirmDelete)
+            {
+                if (!characterObject.ConfirmDelete(LanguageManager.GetString("Message_DeleteArmor",
+                    GlobalOptions.Language)))
+                    return false;
+            }
             DeleteArmor();
             return characterObject.Armor.Remove(this);
         }
